Initial version of the detector
This is a huge commit.  I feel bad about that.
This is also a giant single script, rather than an organized application.
I feel bad about that too.

This represents a starting point in terms of functionality in the name
of being able to launch a usable demo in the next week.  The organization
of the code needs to be redone to be modular and avoid callback hell in
some places.  Overall, however, this commit creates a tool that does the
following:

- Allows Slack teams to sign up for alerts of faces
- Identifies new programs at the TV archive an queues them
- Takes queued programs and downloads the video
- Splits videos into chunks
- Sends chunks of video for analysis at Matroid
- Processes Matroid results and announces them to subscribers

It is a bit hard coded, you can't customize the channel list, and the
detector is specified during server launch.  There is also no sense of
logging or visual output beyond the slack (or watching directories).

All of these things are bad, but this is just an initial demo / prototype
and the overall project is going to undergoe much more significant thinking
so long as I can convince the gods that it's worth spending time on it!