Lowered the processing rate
Over time when there was a lot of backlog the processing rate would
cause faceomatic to start stepping on itself.  This change is likely
temporary -- lowered processing rates should actually be handled at
in terms of a throttle instead (tracking how many are active and
keeping it at that max).